# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request an authenticated private image upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned direct-upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Invalid image metadata
        "401":
          description: Not signed in
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [storage]
      summary: Read an authenticated private object owned by the customer
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "401":
          description: Not signed in
        "403":
          description: Not the object owner
        "404":
          description: Object not found
components:
  schemas:
...
    UploadUrlRequest:
      type: object
      additionalProperties: false
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        size:
          type: integer
          minimum: 1
          maximum: 10485760
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp, image/heic]
    UploadUrlResponse:
      type: object
      additionalProperties: false
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
...